fix: 권한 변경 메소드 user 도메인으로 이동
