+++ src/main/java/com/example/shopping/domain/auth/service/AuthService.java
@@ -2,7 +2,7 @@ package com.example.shopping.domain.auth.service;
 
 import com.example.shopping.config.JwtUtil;
 import com.example.shopping.config.PasswordEncoder;
-import com.example.shopping.domain.auth.dto.request.ChangeUserRoleRequestDto;
+import com.example.shopping.domain.user.dto.ChangeUserRoleRequestDto;
 import com.example.shopping.domain.auth.dto.request.LoginRequestDto;
 import com.example.shopping.domain.auth.dto.request.SignupRequestDto;
 import com.example.shopping.domain.auth.dto.request.WithdrawRequestDto;
@@ -22,6 +22,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Duration;
+import java.util.ArrayList;
+import java.util.List;
 
 @Slf4j
 @Service
@@ -81,39 +83,15 @@ public class AuthService {
             throw new CustomException(ExceptionCode.WRONG_PASSWORD);
         }
 
-        // access token 생성
-        String token = jwtUtil.createToken(user.getId(), user.getEmail(), user.getUserRole());
+        return createNewTokensAndSaveRefresh(user.getId(), user.getEmail(), user.getUserRole());
 
-        // refresh token 생성
-        String refreshToken = jwtUtil.createRefreshToken(user.getId(), user.getEmail(), user.getUserRole());
-
-        // refresh token Redis에 저장
-        // 저장 형태
-        // Key: "refreshToken : 1"
-        // Value: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
-        // TTL: 7일
-        String redisKey = "refreshToken : " + user.getId();
-        redisTemplate.opsForValue().set(redisKey, refreshToken, Duration.ofDays(7));
-
-        return new LoginResponseDto(user.getId(), user.getEmail(), token, refreshToken);
     }
-
 
     // 로그아웃
     public void logout(String bearerAccessToken, Long userId) {
 
         String accessToken = jwtUtil.substringToken(bearerAccessToken);
         addToBlacklist(accessToken);
-
-//        Long remainingTokenTime = jwtUtil.getExpiration(accessToken);
-//
-//        // blacklist token Redis에 저장
-//        // 저장 형태
-//        // Key : "blacklist : eyJhbGci..."
-//        // Value : "logout"
-//        // TTL : 남은 시간
-//        String blackListTokenKey = "blacklist : " + accessToken;
-//        redisTemplate.opsForValue().set(blackListTokenKey, "logout", Duration.ofMillis(remainingTokenTime));
 
         // 저장된 refresh token 삭제
         String refreshTokenKey = "refreshToken : " + userId;
@@ -139,51 +117,46 @@ public class AuthService {
             throw new RuntimeException("유효하지 않은 리프레시 토큰입니다.");
         }
 
-        // 새 토큰 발급, 기존 refresh token 덮어쓰기
-        String newAccessToken = jwtUtil.createToken(userId, email, userRole);
-        String newRefreshToken = jwtUtil.createRefreshToken(userId, email, userRole);
-
-        redisTemplate.opsForValue().set(redisKey, newRefreshToken, Duration.ofDays(7));
-
-        return new LoginResponseDto(userId, email, newAccessToken, newRefreshToken);
+        return createNewTokensAndSaveRefresh(userId, email, userRole);
     }
 
-    public LoginResponseDto changeRole(String bearerAccessToken, ChangeUserRoleRequestDto request){
+    /**
+     * blacklist token Redis에 저장
+     *
+     * 저장 형태
+     * Key : "blacklist : eyJhbGci..."
+     * Value : "logout"
+     * TTL : 남은 시간
+     * @param accessToken
+     */
+    public void addToBlacklist(String accessToken){
 
-        UserRole newUserRole = UserRole.of(request.getUserRole());
+        Long remainingTokenTime = jwtUtil.getExpiration(accessToken);
 
-        String accessToken = jwtUtil.substringToken(bearerAccessToken);
+        String blackListTokenKey = "blacklist : " + accessToken;
+        redisTemplate.opsForValue().set(blackListTokenKey, "logout", Duration.ofMillis(remainingTokenTime));
+    }
 
-        Claims claims = jwtUtil.extractClaims(accessToken);
-        Long userId = Long.parseLong(claims.getSubject());
-        String email = claims.get("email", String.class);
-        UserRole userRole = UserRole.valueOf(claims.get("userRole", String.class));
-
-        if(userRole.equals(newUserRole)){
-            throw new CustomException(ExceptionCode.INVALID_USER_ROLE, "요청한 사용자 권한이 현재 권한과 동일합니다.");
-        }
-
-        addToBlacklist(accessToken);
-
-        String newAccessToken = jwtUtil.createToken(userId, email, newUserRole);
-        String newRefreshToken = jwtUtil.createRefreshToken(userId, email, newUserRole);
+    /**
+     * access, refresh token 생성
+     * refresh token Redis에 저장 or 덮어쓰기
+     *
+     * 저장 형태
+     * Key: "refreshToken : 1"
+     * Value: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
+     * TTL: 7일
+     * @param userId
+     * @param email
+     * @param userRole
+     * @return
+     */
+    public LoginResponseDto createNewTokensAndSaveRefresh(Long userId, String email, UserRole userRole){
+        String newAccessToken = jwtUtil.createToken(userId, email, userRole);
+        String newRefreshToken = jwtUtil.createRefreshToken(userId, email, userRole);
 
         String redisKey = "refreshToken : " + userId;
         redisTemplate.opsForValue().set(redisKey, newRefreshToken, Duration.ofDays(7));
 
         return new LoginResponseDto(userId, email, newAccessToken, newRefreshToken);
-    }
-
-    public void addToBlacklist(String accessToken){
-
-        Long remainingTokenTime = jwtUtil.getExpiration(accessToken);
-
-        // blacklist token Redis에 저장
-        // 저장 형태
-        // Key : "blacklist : eyJhbGci..."
-        // Value : "logout"
-        // TTL : 남은 시간
-        String blackListTokenKey = "blacklist : " + accessToken;
-        redisTemplate.opsForValue().set(blackListTokenKey, "logout", Duration.ofMillis(remainingTokenTime));
     }
 }
